Get rid of compiler warnings.
git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@4139 154be670-7969-47af-9ab0-c6fb378d2295

+++ src/Utils/UnicodeUtils.cpp
@@ -75,7 +75,7 @@ CString CUnicodeUtils::GetUnicode(const CStringA& string)
 #ifdef UNICODE
 std::string CUnicodeUtils::StdGetUTF8(const wide_string& wide)
 {
-	int len = wide.size();
+	int len = (int)wide.size();
 	if (len==0)
 		return std::string();
 	int size = len*4;
@@ -89,7 +89,7 @@ std::string CUnicodeUtils::StdGetUTF8(const wide_string& wide)
 
 wide_string CUnicodeUtils::StdGetUnicode(const std::string& multibyte)
 {
-	int len = multibyte.size();
+	int len = (int)multibyte.size();
 	if (len==0)
 		return wide_string();
 	int size = len*4;
